+++ maia/pytree/node/create.py
@@ -3,10 +3,10 @@ from maia.pytree.meta import begin_api_export, end_api_export
 from maia.pytree import walk
 from .           import access as NA
 
-UNSET = Ellipsis
-
 
 begin_api_export()
+
+UNSET = Ellipsis
 
 def new_node(name:str='Node', label:str='UserDefinedData_t', value:Any=None, children:List[CGNSTree]=[], parent:CGNSTree=None) -> CGNSTree:
   """ Create a new CGNS node